+++ Monitor/src/main/java/racecontrol/extensions/broadcasting/BroadcastingExtension.java
@@ -71,7 +71,7 @@ public class BroadcastingExtension
 
     public void startInstantReplay(float seconds, float duration) {
         LOG.info("Starting instant replay for " + seconds + " seconds");
-        getClient().sendInstantReplayRequest(seconds, duration);
+        getClient().sendInstantReplayRequestSimple(seconds, duration);
     }
 
 }