+++ Monitor/src/main/java/racecontrol/extensions/incidents/IncidentExtension.java
@@ -10,7 +10,6 @@ import racecontrol.client.events.AfterPacketReceived;
 import racecontrol.client.events.BroadcastingEventEvent;
 import racecontrol.eventbus.Event;
 import racecontrol.eventbus.EventBus;
-import racecontrol.eventbus.EventListener;
 import racecontrol.client.extension.AccClientExtension;
 import racecontrol.extensions.GeneralExtentionConfigPanel;
 import racecontrol.extensions.logging.LoggingExtension;
@@ -23,10 +22,12 @@ import racecontrol.extensions.replayoffset.ReplayOffsetExtension;
 import racecontrol.extensions.replayoffset.ReplayStart;
 import racecontrol.client.AccBroadcastingClient;
 import racecontrol.visualisation.gui.LPContainer;
-import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.logging.Logger;
+import racecontrol.client.data.CarInfo;
+import racecontrol.client.events.RealtimeUpdate;
+import racecontrol.client.events.SessionChanged;
 
 /**
  *
@@ -50,11 +51,11 @@ public class IncidentExtension
     /**
      * List of accidents that have happened.
      */
-    private List<IncidentInfo> accidents = new LinkedList<>();
+    private List<IncidentEntry> accidents = new LinkedList<>();
     /**
      * Table model for the incident panel table.
      */
-    private final IncidentTableModel model = new IncidentTableModel();
+    private final IncidentTableModel model;
     /**
      * Flag indicates that the replay offset is known.
      */
@@ -70,9 +71,11 @@ public class IncidentExtension
 
     public IncidentExtension(AccBroadcastingClient client) {
         super(client);
+        this.model = new IncidentTableModel(this);
         this.panel = new IncidentPanel(this);
         loggingExtension = client.getOrCreateExtension(LoggingExtension.class);
         replayOffsetExtension = client.getOrCreateExtension(ReplayOffsetExtension.class);
+
     }
 
     @Override
@@ -91,12 +94,6 @@ public class IncidentExtension
         return model;
     }
 
-    public List<IncidentInfo> getAccidents() {
-        List<IncidentInfo> a = new LinkedList<>(accidents);
-        Collections.reverse(a);
-        return Collections.unmodifiableList(a);
-    }
-
     @Override
     public void onEvent(Event e) {
         if (e instanceof AfterPacketReceived) {
@@ -113,6 +110,26 @@ public class IncidentExtension
             replayTimeKnown = true;
             panel.setReplayOffsetKnown();
             updateAccidentsWithReplayTime();
+        } else if (e instanceof SessionChanged) {
+            List<IncidentEntry> newAccidents = new LinkedList<>();
+            newAccidents.addAll(accidents);
+            switch (((SessionChanged) e).getSessionInfo().getSessionType()) {
+                case PRACTICE:
+                    newAccidents.add(new IncidentEntry(IncidentEntry.Divider.PRACTICE));
+                    break;
+                case QUALIFYING:
+                    newAccidents.add(new IncidentEntry(IncidentEntry.Divider.QUALIFYING));
+                    break;
+                case RACE:
+                    newAccidents.add(new IncidentEntry(IncidentEntry.Divider.RACE));
+                    break;
+            }
+            accidents = newAccidents;
+            model.setAccidents(accidents);
+            panel.invalidate();
+        } else if (e instanceof RealtimeUpdate) {
+            //create a list of all currently connected cars at tell the model about it.
+            model.setConnectedCars(new LinkedList<>(getClient().getModel().getCarsInfo().keySet()));
         }
     }
 
@@ -164,9 +181,9 @@ public class IncidentExtension
     }
 
     private void commitAccident(IncidentInfo a) {
-        List<IncidentInfo> newAccidents = new LinkedList<>();
+        List<IncidentEntry> newAccidents = new LinkedList<>();
         newAccidents.addAll(accidents);
-        newAccidents.add(a);
+        newAccidents.add(new IncidentEntry(a));
         accidents = newAccidents;
         model.setAccidents(accidents);
 
@@ -176,12 +193,21 @@ public class IncidentExtension
 
     private void updateAccidentsWithReplayTime() {
         SessionId currentSessionId = getClient().getSessionId();
-        List<IncidentInfo> newAccidents = new LinkedList<>();
-        for (IncidentInfo incident : accidents) {
-            if (incident.getSessionID().equals(currentSessionId)) {
-                newAccidents.add(incident.withReplayTime(
-                        replayOffsetExtension.getReplayTimeFromSessionTime((int) incident.getSessionEarliestTime())
-                ));
+        List<IncidentEntry> newAccidents = new LinkedList<>();
+        for (IncidentEntry incidentEntry : accidents) {
+            if (incidentEntry.getDividerType() != IncidentEntry.Divider.NONE) {
+                newAccidents.add(incidentEntry);
+            } else {
+                IncidentInfo incident = incidentEntry.getIncident();
+                if (incident.getSessionID().equals(currentSessionId)) {
+                    newAccidents.add(
+                            new IncidentEntry(
+                                    incident.withReplayTime(
+                                            replayOffsetExtension.getReplayTimeFromSessionTime((int) incident.getSessionEarliestTime())
+                                    )
+                            )
+                    );
+                }
             }
         }
         accidents = newAccidents;
@@ -197,6 +223,53 @@ public class IncidentExtension
 
     public void findReplayOffset() {
         replayOffsetExtension.findSessionChange();
+    }
+
+    public void createDummyIncident() {
+        if (getClient().getModel().getCarsInfo().isEmpty()) {
+            return;
+        }
+        LOG.info("Create dummy accident.");
+        int nCars = (int) Math.floor(Math.random() * Math.min(6, getClient().getModel().getCarsInfo().size()) + 1);
+        float sessionTime = getClient().getModel().getSessionInfo().getSessionTime();
+        IncidentInfo incident = new IncidentInfo(
+                sessionTime,
+                replayOffsetExtension.getReplayTimeFromSessionTime((int) sessionTime),
+                getClient().getSessionId());
+        for (int i = 0; i < nCars; i++) {
+            incident = incident.addCar(
+                    sessionTime,
+                    getRandomCar(),
+                    0);
+        }
+        commitAccident(incident);
+    }
+
+    private CarInfo getRandomCar() {
+        int r = (int) Math.floor(Math.random() * getClient().getModel().getCarsInfo().size());
+        int i = 0;
+        for (CarInfo car : getClient().getModel().getCarsInfo().values()) {
+            if (i++ == r) {
+                return car;
+            }
+        }
+        return null;
+    }
+
+    public void focusOnCar(int carId) {
+        LOG.info("Setting focus on car " + carId);
+        getClient().sendChangeFocusRequest(carId);
+    }
+
+    public void startAccidentReplay(IncidentInfo incident, int seconds) {
+        LOG.info("Starting instant replay for incident for " + seconds);
+        getClient().sendInstantReplayRequestWithCamera(
+                incident.getSessionEarliestTime() - seconds * 1000f - 5000,
+                seconds + 5,
+                getClient().getModel().getSessionInfo().getFocusedCarIndex(),
+                getClient().getModel().getSessionInfo().getActiveCameraSet(),
+                getClient().getModel().getSessionInfo().getActiveCamera()
+        );
     }
 
 }