+++ Monitor/src/main/java/racecontrol/extensions/incidents/IncidentTableModel.java
@@ -5,6 +5,7 @@
  */
 package racecontrol.extensions.incidents;
 
+import java.util.Arrays;
 import racecontrol.client.data.CarInfo;
 import racecontrol.utility.TimeUtils;
 import racecontrol.visualisation.LookAndFeel;
@@ -18,9 +19,17 @@ import racecontrol.visualisation.gui.LPTable;
 import racecontrol.visualisation.gui.TableModel;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.logging.Logger;
+import java.util.stream.Collectors;
 import processing.core.PApplet;
 import static processing.core.PConstants.CENTER;
 import static processing.core.PConstants.CLOSE;
+import static racecontrol.visualisation.LookAndFeel.COLOR_DARK_RED;
+import static racecontrol.visualisation.LookAndFeel.COLOR_GRAY;
+import static racecontrol.visualisation.LookAndFeel.COLOR_PRACTICE;
+import static racecontrol.visualisation.LookAndFeel.COLOR_QUALIFYING;
+import static racecontrol.visualisation.LookAndFeel.COLOR_RACE;
+import static racecontrol.visualisation.LookAndFeel.COLOR_RED;
 
 /**
  *
@@ -28,19 +37,43 @@ import static processing.core.PConstants.CLOSE;
  */
 public class IncidentTableModel extends TableModel {
 
-    private List<IncidentInfo> accidents = new LinkedList<>();
+    /**
+     * This class's logger.
+     */
+    public static final Logger LOG = Logger.getLogger(IncidentTableModel.class.getName());
+    /**
+     * Maximum cars per row.
+     */
+    public static final int MAX_CARS_PER_ROW = 4;
+    /**
+     * List of incident entries to display.
+     */
+    private List<IncidentEntry> incidents = new LinkedList<>();
+    /**
+     * List of cars that are currently connected.
+     */
+    private List<Integer> currentlyConnectedCars;
+    /**
+     * Reference to the base extension object.
+     */
+    private final IncidentExtension extension;
+
+    public IncidentTableModel(IncidentExtension extension) {
+        this.extension = extension;
+    }
 
     @Override
     public int getRowCount() {
-        return accidents.size();
+        return incidents.stream()
+                .collect(Collectors.summingInt(entry -> entry.getRows()));
     }
 
     @Override
     public LPTableColumn[] getColumns() {
         return new LPTableColumn[]{
             new LPTableColumn("#")
-            .setMaxWidth(LookAndFeel.LINE_HEIGHT)
-            .setMinWidth(LookAndFeel.LINE_HEIGHT),
+            .setMaxWidth(LINE_HEIGHT)
+            .setMinWidth(LINE_HEIGHT),
             new LPTableColumn("Session Time")
             .setMaxWidth(200)
             .setMinWidth(200),
@@ -48,16 +81,53 @@ public class IncidentTableModel extends TableModel {
             .setMaxWidth(200)
             .setMinWidth(200),
             new LPTableColumn("Cars Involved")
-            .setCellRenderer(carsRenderer)
+            .setMinWidth(LINE_HEIGHT * 1.25f * MAX_CARS_PER_ROW)
+            .setMaxWidth(LINE_HEIGHT * 1.25f * MAX_CARS_PER_ROW)
+            .setCellRenderer(carsRenderer),
+            new LPTableColumn("Replay")
+            .setMinWidth(LINE_HEIGHT * 6.5f)
+            .setMaxWidth(LINE_HEIGHT * 6.5f)
+            .setCellRenderer(replayButtonRenderer),
+            new LPTableColumn("")
+            .setMaxWidth(0)
+            .setMinWidth(0)
+            .setCellRenderer(dividerRender)
         };
     }
 
     @Override
     public Object getValueAt(int column, int row) {
-        IncidentInfo a = accidents.get(accidents.size() - row - 1);
+        IncidentEntry entry = null;
+        int subRow = 0;
+        //find correct Entry
+        int rowCount = 0;
+        for (int i = 0; i < incidents.size(); i++) {
+            entry = incidents.get(i);
+            rowCount += entry.getRows();
+            subRow = entry.getRows() - (rowCount - row);
+            if (rowCount > row) {
+                break;
+            }
+        }
+
+        if (entry == null) {
+            return null;
+        }
+        if (subRow > 0 && column != 3) {
+            return null;
+        }
+        if (entry.getDividerType() != IncidentEntry.Divider.NONE) {
+            if (column == 5) {
+                return entry.getDividerType();
+            } else {
+                return null;
+            }
+        }
+
+        IncidentInfo a = entry.getIncident();
         switch (column) {
             case 0:
-                return String.valueOf(accidents.size() - row - 1);
+                return String.valueOf(row);
             case 1:
                 return TimeUtils.asDuration(a.getSessionEarliestTime());
             case 2:
@@ -66,9 +136,55 @@ public class IncidentTableModel extends TableModel {
                 }
                 return "-";
             case 3:
-                return a.getCars();
+                int lower = MAX_CARS_PER_ROW * subRow;
+                int upper = Math.min(MAX_CARS_PER_ROW * (subRow + 1), a.getCars().size());
+                return a.getCars().subList(lower, upper);
+            case 4:
+                return "-";
         }
-        return "-";
+        return null;
+    }
+
+    @Override
+    public void onClick(int column, int row, int mouseX, int mouseY) {
+        IncidentEntry entry = null;
+        int subRow = 0;
+        //find correct Entry
+        int rowCount = 0;
+        for (int i = 0; i < incidents.size(); i++) {
+            entry = incidents.get(i);
+            rowCount += entry.getRows();
+            subRow = entry.getRows() - (rowCount - row);
+            if (rowCount > row) {
+                break;
+            }
+        }
+        if (entry == null) {
+            return;
+        }
+
+        if (column == 3) {
+            //Car column clicked
+            int index = (int) (mouseX / (LINE_HEIGHT * 1.25f)) + MAX_CARS_PER_ROW * subRow;
+            if (index >= entry.getIncident().getCars().size()) {
+                return;
+            }
+            extension.focusOnCar(entry.getIncident().getCars().get(index).getCarId());
+
+        } else if (column == 4 && subRow == 0) {
+            //replay column clicked on the first sub row.
+            int button = (int) Math.floor((mouseX - LINE_HEIGHT) / (LINE_HEIGHT * 1.5f));
+            if (button > 2 || button < 0) {
+                return;
+            }
+            int seconds = 20;
+            if (button == 1) {
+                seconds = 10;
+            } else if (button == 2) {
+                seconds = 5;
+            }
+            extension.startAccidentReplay(entry.getIncident(), seconds);
+        }
     }
 
     private final LPTable.CellRenderer carsRenderer = (
@@ -101,6 +217,20 @@ public class IncidentTableModel extends TableModel {
             applet.fill(background_color);
             applet.rect(x + 1, 1, w - 2, context.height - 2);
 
+            if (currentlyConnectedCars.contains(car.getCarId())) {
+                //draw outline if the mouse if over this car
+                if (context.isMouseOverColumn
+                        && context.isMouseOverRow
+                        && context.mouseX > x
+                        && context.mouseX < x + w) {
+                    applet.fill(COLOR_DARK_RED);
+                    applet.rect(x + 1, 1, w - 2, 3);
+                    applet.rect(x + 1, context.height - 1, w - 2, -3);
+                    applet.rect(x + 1, 1, 3, context.height - 2);
+                    applet.rect(x + w - 1, 1, -3, context.height - 2);
+                }
+            }
+
             //render GT4 / Cup / Super trofeo corners.
             CarType type = getCarType(car.getCarModelType());
             if (type != CarType.GT3) {
@@ -128,12 +258,67 @@ public class IncidentTableModel extends TableModel {
             applet.textAlign(CENTER, CENTER);
             applet.textFont(LookAndFeel.fontMedium());
             applet.text(String.valueOf(carNumber), x + w / 2, context.height / 2f);
+
+            //Draw car number darker if this car is not connected.
+            if (!currentlyConnectedCars.contains(car.getCarId())) {
+                applet.fill(0, 0, 0, 150);
+                applet.rect(x + 1, 1, w - 2, context.height - 2);
+            }
+
             x += w;
         }
     };
 
-    public void setAccidents(List<IncidentInfo> accidents) {
-        this.accidents = accidents;
+    private final LPTable.CellRenderer replayButtonRenderer = (
+            PApplet applet,
+            LPTable.RenderContext context) -> {
+        int width = (int) (LINE_HEIGHT * 1.5f);
+        int x = LINE_HEIGHT;
+        for (String t : Arrays.asList("-20s", "-10s", "-5s")) {
+            if (context.isMouseOverColumn
+                    && context.isMouseOverRow
+                    && context.mouseX > x
+                    && context.mouseX < x + width) {
+                applet.fill(COLOR_RED);
+            } else {
+                applet.fill(COLOR_GRAY);
+            }
+            applet.rect(x + 1, 1, width - 2, context.height - 2);
+            applet.fill(COLOR_WHITE);
+            applet.textAlign(CENTER, CENTER);
+            applet.text(t, x + width / 2f, context.height / 2);
+            x += width;
+        }
+    };
+
+    private final LPTable.CellRenderer dividerRender = (
+            PApplet applet,
+            LPTable.RenderContext context) -> {
+        switch ((IncidentEntry.Divider) context.object) {
+            case PRACTICE:
+                applet.fill(COLOR_PRACTICE);
+                break;
+            case QUALIFYING:
+                applet.fill(COLOR_QUALIFYING);
+                break;
+            case RACE:
+                applet.fill(COLOR_RACE);
+                break;
+        }
+        applet.rect(-context.tablePosX + 1, 1, context.tableWidth - 2, context.height - 2);
+        applet.fill(0);
+        applet.textAlign(CENTER, CENTER);
+        applet.text(((IncidentEntry.Divider) context.object).name(),
+                -context.tablePosX + context.tableWidth / 2f,
+                context.height / 2f);
+    };
+
+    public void setAccidents(List<IncidentEntry> incidents) {
+        this.incidents = incidents;
+    }
+
+    public void setConnectedCars(List<Integer> connectedCars) {
+        this.currentlyConnectedCars = connectedCars;
     }
 
     private CarType getCarType(byte carModelId) {
@@ -156,6 +341,7 @@ public class IncidentTableModel extends TableModel {
                 return CarType.GT4;
             default:
                 return CarType.GT3;
+
         }
     }