+++ ChangeLog.md
@@ -1,6 +1,8 @@
 # v1.2.0
 
 * Add a broadcasting panel for easier control of the game for live broadcasts.
+* You can now start an instant replay for an incident either 20s, 10s, or 5s before the contact happened by clicking on the corresponding button in the incident panel.
+* You can now change the focused car by clicking on the car number in the incident panel.
 * Code refactoring.
 
 Bug fixes: